TUI nits: extract shared catalog-card helpers (DRY across card/grid/detail)
Move the byte-identical name-truncation, spec/status pills, fit-level color map,
and verbose fit-pill renderer into catalog_card_shared; model_card, model_grid,
and catalog_detail import them instead of each carrying a copy. Also fixes the
grid drawer's stale 'won't -X GB' label (now shares the fixed renderer).
